functions.cpp

+++ OOP WITH C++/functions.cpp
@@ -1,11 +1,11 @@
 #include <iostream>
 using namespace std;
 
-int average(int a, int b){
+double average(int a, int b){
     return (a + b)/2;
 };
 int main(){
-    int result = average(43,176);
+    double result = average(43,176);
     cout<< "the average is : "<<result <<endl;
 
     return 0;
